Create smartCode.c
The idea is to develop a cheap  solution  which will help in crop monitoring and smarrt Farming using IOT, the buzz word in the field of Technology.

Main idea is to digitize farming and agricultural activities so that the farmers can check on the requirements of the crops and predict their growth in real time accurately. This concept will surely accelerate their business to reach new heights and also be more profitable. The implementation of our project largely depends upon the awareness among farmers.

Main idea is to introduce the latest technology into the agriculture business and better crop production by collecting real-time status of crop and informing the farmers about it.

The current solution is a basic one implemented in Arduino Uno which will be collecting data from the soil about eh moisture content of the soil, humidity and other factors responsible for crop production. 

Drones could also have been used for collecting the data but they are a costly solution